backend/pyproject.toml
[project]
name = "tee-mo-backend"
version = "0.1.0"
description = "Tee-Mo backend API — FastAPI-based AI assistant for Slack workspaces"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi[standard]==0.135.3",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.30.0",
    "supabase==2.28.3",
    "pydantic-ai[openai,anthropic,google]==1.79.0",
    "cryptography==46.0.7",
    "PyJWT==2.12.1",
    "bcrypt==5.0.0",
    "slack-bolt==1.28.0",
    "google-api-python-client==2.194.0",
    "google-auth==2.49.2",
    "pymupdf4llm",
    "python-docx",
    "openpyxl",
    "python-json-logger>=2.0.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
